+++ backend/pydantic_schemas/kidney_schema.py
@@ -6,9 +6,9 @@ import datetime
 # for kidney test table
 class Kidney(BaseModel):
     k_bu : int
-    k_cc : int
+    k_cc : float
     k_ss : int
-    k_sp : int
+    k_sp : float
     k_sc : int
-    k_sca : int
-    k_ur : int
+    k_sca : float
+    k_ur : float
